API_INTEGRACAO/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
pymysql